pyproject.toml
[project]
name = "bohe-api-auto-sign"
version = "0.0.0"
description = "Auto sign bohe api"
authors = [
    { name = "Sn0wo2" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "linux-do-connect-token (==0.0.2b2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
